2010.11

+++ test/Bayesian.cc
@@ -12,7 +12,7 @@ using std::cout;
 using std::endl;
 using namespace lands;
 void processParameters(int argc, const char* argv[]);
-int debug=0; int nexps=20000; double s=0;  double b=0; double s_err = 0; double b_err = 0; int d=0;
+int debug=0; int nexps=20000; double s=0;  double b=0; double s_err = 0; double b_err = 0; double d=0;
 int seed =1234; int pdftypeEs = 1; int pdftypeEb = 1; int EsEb_correlated = 0; int calcExpectedMeanLimit=0; //int calcSignificance=0;	 
 int main(int argc, const char* argv[]){
 	processParameters(argc, argv);
@@ -77,7 +77,7 @@ void processParameters(int argc, const char* argv[]){
 		s_err=atof( argv[2] );
 		b=atof( argv[3] );
 		b_err=atof( argv[4] );
-		d=atoi( argv[5] );
+		d=atof( argv[5] );
 		npar++;
 		if(argc>=npar+1){
 			calcExpectedMeanLimit=atoi(argv[npar]);
